requirements.txt
beautifulsoup4==4.11.1
jsonlines==3.0.0
python-docx==0.8.11
